pyflow_server/requirements.txt
# PyFlow CLI Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies for video downloading
yt-dlp>=2024.1.0

# GUI Framework
customtkinter>=5.2.0
pystray>=0.19.0
Pillow>=10.0.0

# CLI & Rich
rich>=13.0.0
pyperclip>=1.8.2

# Web server for Chrome Extension communication
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Optional modern ttk theming for GUI polish
ttkbootstrap>=1.12.0

# Additional utilities
httpx>=0.24.0
pydantic>=2.0.0
